Poprawa wyświetlania błędów, sprawdzania poprawności danych przesyłanych w formularzu, niedopuszczanie wartości ujemnych

+++ nutrimngr/db/migrate/20200418144733_add_recipe_products_table.rb
@@ -2,7 +2,7 @@ class AddRecipeProductsTable < ActiveRecord::Migration[5.2]
   def change
 	create_table :RecipeProducts do |t|
 	 t.integer :IDP
-	 t.float :Weight, precision: 6, scale: 2, null: true
+	 t.numeric :Weight, precision: 6, scale: 2, null: true
 	end
   end
 end
